+++ core-ng/src/main/java/core/framework/http/HTTPClient.java
@@ -9,4 +9,6 @@ public interface HTTPClient {
     }
 
     HTTPResponse execute(HTTPRequest request);
+
+    EventSource sse(HTTPRequest request);
 }