+++ core-ng/src/main/java/core/framework/internal/web/sse/ServerSentEventBuilder.java
@@ -21,8 +21,8 @@ class ServerSentEventBuilder<T> {
 
     String build(String id, String data) {
         var builder = new StringBuilder(data.length() + 7 + (id == null ? 0 : id.length() + 4));
-        if (id != null) builder.append("id:").append(id).append('\n');
-        builder.append("data:").append(data).append("\n\n");
+        if (id != null) builder.append("id: ").append(id).append('\n');
+        builder.append("data: ").append(data).append("\n\n");
 
         return builder.toString();
     }